fix cleanup and run on windows

+++ assets/shaders/anim_tbn.vert
@@ -4,7 +4,7 @@ layout (location=0) in vec3 aPos;
 layout (location=1) in vec3 aNormal;
 layout (location=2) in vec2 aUV;
 layout (location=3) in vec3 aTangent;
-layout (location=4) in uvec4 aJoints;
+layout (location=4) in vec4 aJoints;
 layout (location=5) in vec4 aWeights;
 
 layout (std140) uniform CameraBlock
